refactor responsive view options
