+++ src/Responsive.Core/src/Abstractions/IResponsiveOptions.cs
@@ -1,6 +1,8 @@
 // Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
+using Wangkanai.Detection;
+
 namespace Wangkanai.Responsive
 {
     /// <summary>
@@ -12,5 +14,24 @@ namespace Wangkanai.Responsive
         /// Get or set the option to configure the default view.
         /// </summary>
         IResponsiveViewOptions View { get; set; }
+    }
+
+    /// <summary>
+    /// The IResponsiveViewOptions class is the View container for all configuration settings of Responsive.
+    /// </summary>
+    public interface IResponsiveViewOptions
+    {
+        /// <summary>
+        /// Gets or sets a value that determines the default view for Mobile
+        /// </summary>
+        DeviceType DefaultMobile { get; set; }
+        /// <summary>
+        /// Gets or sets a value that determines the default view for Tablet
+        /// </summary>
+        DeviceType DefaultTablet { get; set; }
+        /// <summary>
+        /// Gets or sets a value that determines the default view for Desktop
+        /// </summary>
+        DeviceType DefaultDesktop { get; set; }
     }
 }
