--- src/Responsive.Core/src/Abstractions/IResponsiveResolver.cs
@@ -1,12 +1,0 @@
-// Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
-// The Apache v2. See License.txt in the project root for license information.
-
-using Wangkanai.Detection;
-
-namespace Wangkanai.Responsive
-{
-    public interface IResponsiveResolver
-    {
-        IDevice Device { get; }
-    }
-}
